requirements.txt
accelerate==1.12.0
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
av==16.1.0
bce-python-sdk==0.9.60
certifi==2026.2.25
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cuda-python==13.0.3
cycler==0.12.1
diffusers @ git+https://github.com/huggingface/diffusers.git@47e8faf3b99ecd1f194161d7647a44e263f726a3
filelock==3.24.3
fonttools==4.61.1
fsspec==2026.2.0
future==1.0.0
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
imagesize==1.4.1
importlib_metadata==8.7.1
Jinja2==3.1.6
joblib==1.5.3
kintera==1.3.1
kiwisolver==1.4.9
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.34.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
nvidia-cublas==13.0.2.14
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cccl==13.0.85
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime==13.0.88
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile==1.15.1.6
nvidia-cufile-cu12==1.11.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-cusparselt-cu13==0.8.1
nvidia-nccl-cu12==2.27.5
nvidia-nccl-cu13==2.28.3
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.6.77
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
packaging @ file:///home/task_176104874243446/conda-bld/packaging_1761049080023/work
paddle==1.2.7
paddleocr==3.4.0
paddlepaddle-gpu==3.3.0
paddlex==3.4.2
pandas==3.0.1
peft==0.18.1
pillow==12.1.1
prettytable==3.17.0
protobuf==6.33.4
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydisort==1.5.0
Pygments==2.19.2
pyharp==1.9.0
pynndescent==0.6.0
pyparsing==3.3.2
pypdfium2==5.5.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qwen-vl-utils==0.0.14
regex==2026.2.28
requests==2.32.5
rich==14.3.3
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==80.10.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
snapy==1.3.4
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0+cu126
torchvision==0.25.0+cu126
tqdm==4.67.3
transformers==5.2.0
triton==3.6.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
umap-learn==0.5.11
urllib3==2.6.3
wcwidth==0.6.0
wheel==0.46.3
zipp==3.23.0
